+++ src/app/core/+store/app.state.ts
@@ -1,0 +1,5 @@
+import { ProductsState } from '.';
+
+export interface AppState {
+  products: ProductsState;
+}