+++ src/app/core/guards/product.guard.ts
@@ -2,23 +2,23 @@ import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, RouterStateSnapshot, Resolve } from '@angular/router';
 import { Observable } from 'rxjs';
 import { IProduct } from 'src/app/products/interfaces/iproduct';
-import { ProductService } from 'src/app/products/services/product.service';
-import { Product } from 'src/app/products/models/product';
+import { AppState, SetProduct } from '../+store';
+import { Store, select } from '@ngrx/store';
+import { getProductByUrl } from '../+store/products/products.selectors';
+import { tap, map, take } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductResolveGuard implements Resolve<IProduct> {
 
-  constructor(private productService: ProductService) {
+  constructor(private store: Store<AppState>) {
   }
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): IProduct | Observable<IProduct> | Promise<IProduct> {
-    const id = route.params.id;
-    if (!id) {
-      return new Product();
-    } else {
-      return { ...this.productService.getProduct(+id) };
-    }
+    return this.store.pipe(
+      select(getProductByUrl),
+      tap(product => this.store.dispatch(new SetProduct(product))),
+      take(1));
   }
 }